requirements/requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --extra=dev --extra=test --output-file=requirements/requirements-dev.txt pyproject.toml
#
asgiref==3.8.1
    # via
    #   django
    #   django-ansible-base
astor==0.8.1
    # via flynt
black==24.10.0
    # via pattern_service (pyproject.toml)
build==1.2.2.post1
    # via pip-tools
cachetools==6.1.0
    # via tox
certifi==2025.8.3
    # via requests
cffi==1.17.1
    # via cryptography
chardet==5.2.0
    # via tox
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   black
    #   pip-tools
colorama==0.4.6
    # via tox
cryptography==45.0.4
    # via django-ansible-base
dispatcherd==2025.5.21
    # via pattern_service (pyproject.toml)
distlib==0.3.9
    # via virtualenv
django==4.2.23
    # via
    #   django-ansible-base
    #   django-crum
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework
django-ansible-base[resource-registry]==2025.5.8
    # via pattern_service (pyproject.toml)
django-crum==0.7.9
    # via django-ansible-base
django-stubs==5.2.1
    # via pattern_service (pyproject.toml)
django-stubs-ext==5.2.1
    # via django-stubs
djangorestframework==3.16.0
    # via django-ansible-base
dynaconf==3.2.11
    # via django-ansible-base
filelock==3.18.0
    # via
    #   tox
    #   virtualenv
flake8==6.1.0
    # via pattern_service (pyproject.toml)
flynt==1.0.2
    # via pattern_service (pyproject.toml)
idna==3.10
    # via requests
inflection==0.5.1
    # via django-ansible-base
iniconfig==2.1.0
    # via pytest
isort==5.13.2
    # via pattern_service (pyproject.toml)
mccabe==0.7.0
    # via flake8
mypy==1.16.0
    # via pattern_service (pyproject.toml)
mypy-extensions==1.1.0
    # via
    #   black
    #   mypy
packaging==25.0
    # via
    #   black
    #   build
    #   pyproject-api
    #   pytest
    #   tox
pathspec==0.12.1
    # via
    #   black
    #   mypy
pip-tools==7.4.1
    # via pattern_service (pyproject.toml)
platformdirs==4.3.8
    # via
    #   black
    #   tox
    #   virtualenv
pluggy==1.6.0
    # via
    #   pytest
    #   tox
psycopg==3.2.9
    # via pattern_service (pyproject.toml)
psycopg-binary==3.2.9
    # via pattern_service (pyproject.toml)
pycodestyle==2.11.1
    # via flake8
pycparser==2.22
    # via cffi
pyflakes==3.1.0
    # via flake8
pygments==2.19.2
    # via pytest
pyjwt==2.10.1
    # via django-ansible-base
pyproject-api==1.9.1
    # via tox
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==8.4.1
    # via pytest-django
pytest-django==4.11.1
    # via pattern_service (pyproject.toml)
python-dotenv==1.1.1
    # via pattern_service (pyproject.toml)
pyyaml==6.0.2
    # via dispatcherd
requests==2.32.4
    # via
    #   django-ansible-base
    #   pattern_service (pyproject.toml)
sqlparse==0.5.3
    # via
    #   django
    #   django-ansible-base
tox==4.27.0
    # via pattern_service (pyproject.toml)
types-pyyaml==6.0.12.20250516
    # via django-stubs
types-requests==2.32.4.20250611
    # via pattern_service (pyproject.toml)
typing-extensions==4.13.2
    # via
    #   django-stubs
    #   django-stubs-ext
    #   mypy
urllib3==2.5.0
    # via
    #   django-ansible-base
    #   requests
    #   types-requests
virtualenv==20.31.2
    # via tox
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
